pyproject.toml
[tool.poetry]
name = "desktop"
version = "0.1.0"
description = ""
authors = ["那須優斗 <youdounaxu@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.3"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
